refactor: resolve ESLint function and file length limits after code formatting
